refactor layouts to have interview and dashboard separate layouts
fixes hydration error. currently is repetitive with two nextintlclientproviders so may need a diff approach
