+++ app/[locale]/(dashboard)/layout.tsx
@@ -1,18 +1,37 @@
 import ResponsiveContainer from '~/components/layout/ResponsiveContainer';
-
 import LanguageSwitcher from '../../_components/LanguageSwitcher';
+import { NextIntlClientProvider } from 'next-intl';
+import { Inter } from 'next/font/google';
+import { getMessages } from 'next-intl/server';
+import { type Metadata } from 'next';
+import '~/app/globals.scss';
 
-export default function DashboardLayout({
+const inter = Inter({ subsets: ['latin'] });
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
+
+export default async function DashboardLayout({
   children,
+  params: { locale },
 }: {
   children: React.ReactNode;
+  params: { locale: string };
 }) {
+  // Providing all messages to the client
+  // side is the easiest way to get started
+  const messages = await getMessages();
   return (
-    <div>
-      <div className="m-2">
-        <LanguageSwitcher />
-      </div>
-      <ResponsiveContainer>{children}</ResponsiveContainer>
-    </div>
+    <html lang={locale}>
+      <body className={inter.className}>
+        <NextIntlClientProvider messages={messages}>
+          <div className="m-2">
+            <LanguageSwitcher />
+          </div>
+          <ResponsiveContainer>{children}</ResponsiveContainer>
+        </NextIntlClientProvider>
+      </body>
+    </html>
   );
 }
