--- app/[locale]/layout.tsx
@@ -1,34 +1,0 @@
-import { NextIntlClientProvider } from 'next-intl';
-import { Inter } from 'next/font/google';
-import { getMessages } from 'next-intl/server';
-import { type Metadata } from 'next';
-import '../globals.scss';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-
-export default async function LocaleLayout({
-  children,
-  params: { locale },
-}: {
-  children: React.ReactNode;
-  params: { locale: string };
-}) {
-  // Providing all messages to the client
-  // side is the easiest way to get started
-  const messages = await getMessages();
-
-  return (
-    <html lang={locale}>
-      <body className={inter.className}>
-        <NextIntlClientProvider messages={messages}>
-          {children}
-        </NextIntlClientProvider>
-      </body>
-    </html>
-  );
-}
